executing based on the PATH env variable working properly
